+++ eth/backend.go
@@ -615,6 +615,15 @@ func New(ctx context.Context, stack *node.Node, config *ethconfig.Config, logger
 				EventFetcher: heimdallClient,
 			})
 
+			if err := heimdallStore.Milestones().Prepare(ctx); err != nil {
+				return nil, err
+			}
+
+			_, err := heimdallStore.Milestones().DeleteFromBlockNum(ctx, 0)
+			if err != nil {
+				return nil, err
+			}
+
 			heimdallService = heimdall.NewService(heimdall.ServiceConfig{
 				Store:     heimdallStore,
 				BorConfig: borConfig,
@@ -1885,16 +1894,9 @@ func setBorDefaultMinerGasPrice(chainConfig *chain.Config, config *ethconfig.Con
 }
 
 func setDefaultMinerGasLimit(chainConfig *chain.Config, config *ethconfig.Config, logger log.Logger) {
-	if chainConfig.Bor != nil {
-		if config.Miner.GasLimit == nil {
-			gasLimit := ethconfig.BorDefaultMinerGasLimit
-			config.Miner.GasLimit = &gasLimit
-		}
-	} else {
-		if config.Miner.GasLimit == nil {
-			gasLimit := ethconfig.DefaultMinerGasLimit
-			config.Miner.GasLimit = &gasLimit
-		}
+	if config.Miner.GasLimit == nil {
+		gasLimit := ethconfig.DefaultMinerGasLimitByChain(config)
+		config.Miner.GasLimit = &gasLimit
 	}
 }