+++ turbo/execution/eth1/ethereum_execution.go
@@ -284,7 +284,7 @@ func (e *EthereumExecutionModule) ValidateChain(ctx context.Context, req *execut
 		return nil, err
 	}
 
-	tx, err := e.db.BeginRw(ctx)
+	tx, err := e.db.BeginRwNosync(ctx)
 	if err != nil {
 		return nil, err
 	}
@@ -296,7 +296,7 @@ func (e *EthereumExecutionModule) ValidateChain(ctx context.Context, req *execut
 	}
 	// Throw away the tx and start a new one (do not persist changes to the canonical chain)
 	tx.Rollback()
-	tx, err = e.db.BeginRw(ctx)
+	tx, err = e.db.BeginRwNosync(ctx)
 	if err != nil {
 		return nil, err
 	}